Rewrite Label for consistency; expand and fix tests.
Label is no longer a namedtuple.  It enforces that (content_id1, subtopic_id1)
is less than or equal to (content_id2, subtopic_id2), swapping in the
constructor if needed, which simplifies other code.  Make equality and
ordering work correctly in a way that is obvious without knowing the
database storage (for instance, two Labels that are equal except that
one is a positive label and one is a negative label are no longer ==).

LabelStore.everything() is much expanded, and tries much harder to
return only one Label even though many things will be recorded in the
database twice.  The kvlayer encoding/decoding moves from Label to
LabelStore.

Start in on tests for the CLI path; this is not done yet.